requirements.txt
openai==1.14.0
pandas==2.2.1
python-dotenv==1.0.1
google-api-python-client==2.122.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow==10.2.0
pdf2image==1.17.0
